// backend/auth-service/src/proto/generated/doctor.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.8
//   protoc               v3.21.12
// source: doctor.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "doctor";

export interface DoctorProfileReq {
  doctorPrimaryKey: number;
  doctorId: string;
  email: string;
  mobile: string;
  firstName: string;
  middleName?: string | undefined;
  lastName: string;
}

export interface DoctorProfileRes {
  doctorId: string;
}

export const DOCTOR_PACKAGE_NAME = "doctor";

export interface DoctorServiceClient {
  createDoctorProfile(request: DoctorProfileReq): Observable<DoctorProfileRes>;
}

export interface DoctorServiceController {
  createDoctorProfile(
    request: DoctorProfileReq,
  ): Promise<DoctorProfileRes> | Observable<DoctorProfileRes> | DoctorProfileRes;
}

export function DoctorServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["createDoctorProfile"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("DoctorService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("DoctorService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const DOCTOR_SERVICE_NAME = "DoctorService";