// backend/auth-service/src/proto/generated/patient.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.8
//   protoc               v3.21.12
// source: patient.proto

/* eslint-disable */
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "patient";

export interface PatientProfileReq {
  patientPrimaryKey: number;
  patientId: string;
  email: string;
  mobile: string;
  firstName: string;
  middleName?: string | undefined;
  lastName: string;
}

export interface PatientProfileRes {
  patientId: string;
}

export const PATIENT_PACKAGE_NAME = "patient";

export interface PatientServiceClient {
  createPatient(request: PatientProfileReq): Observable<PatientProfileRes>;
}

export interface PatientServiceController {
  createPatient(
    request: PatientProfileReq,
  ): Promise<PatientProfileRes> | Observable<PatientProfileRes> | PatientProfileRes;
}

export function PatientServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = ["createPatient"];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("PatientService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("PatientService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const PATIENT_SERVICE_NAME = "PatientService";